JWT Update

+++ E-Commerce.DAL/Entities/OrderAggregate/OrderItem.cs
@@ -13,13 +13,13 @@ namespace E_Commerce.DAL.Entities.OrderAggregate
         {
         }
 
-        public OrderItem(ProductItemOrdered itemOrdered, decimal price, int quantity)
+        public OrderItem(ProductItemOrdered itemOrdered, double price, int quantity)
         {
             ItemOrdered = itemOrdered;
             Price = price;
             Quantity = quantity;
         }
-        public decimal Price { get; set; }
+        public double Price { get; set; }
         public int Quantity { get; set; }
         public ProductItemOrdered ItemOrdered { get; set; }
